Moved the new save/load methods out to an extension class specifically for the batched executor.
